reset: starfive: Use 32bit I/O on 32bit registers

+++ drivers/reset/starfive/reset-starfive-jh7100.h
@@ -9,7 +9,7 @@
 #include <linux/platform_device.h>
 
 int starfive_jh7100_reset_init(struct platform_device *pdev,
-			       const u64 *asserted,
+			       const u32 *asserted,
 			       unsigned int status_offset,
 			       unsigned int nr_resets);
